ml/functions/shows-knn-serve/requirements.txt
functions-framework
pandas
numpy
joblib
Flask
gcsfs
scikit-learn
google-cloud-storage
google-cloud-aiplatform